[svn-r29138] Correct int to long issues

+++ java/examples/datatypes/H5Ex_T_Opaque.java
@@ -144,7 +144,7 @@ public class H5Ex_T_Opaque {
         long datatype_id = -1;
         long dataspace_id = -1;
         long dataset_id = -1;
-        int type_len = -1;
+        long type_len = -1;
         long[] dims = { DIM0 };
         byte[] dset_data;
         String tag_name = null;
@@ -197,7 +197,7 @@ public class H5Ex_T_Opaque {
         }
 
         // Allocate buffer.
-        dset_data = new byte[(int) dims[0] * type_len];
+        dset_data = new byte[(int) (dims[0] * type_len)];
 
         // Read data.
         try {
@@ -214,7 +214,7 @@ public class H5Ex_T_Opaque {
         for (int indx = 0; indx < dims[0]; indx++) {
             System.out.print(DATASETNAME + "[" + indx + "]: ");
             for (int jndx = 0; jndx < type_len; jndx++) {
-                char temp = (char) dset_data[jndx + indx * type_len];
+                char temp = (char) dset_data[jndx + indx * (int)type_len];
                 System.out.print(temp);
             }
             System.out.println("");
